dashboard/requirements.txt
flask==2.2.3
flask-cors==3.0.10
sqlalchemy==1.4.46
Pillow>=10.0.0
piexif==1.1.3
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.2
six==1.16.0
python-dateutil==2.8.2
pytz==2023.3
watchdog==3.0.0
